backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pg8000
sqlparse
Nominatim
geopy
spacy